pull CPU usage test out of manager
